feat: ADD cache for local pull

+++ pull_to_local.py
@@ -1,3 +1,4 @@
+import json
 import re
 from index import ConfigYaml
 import os
@@ -5,11 +6,18 @@ import os
 import_command = 'microk8s ctr image import '
 
 if __name__ == '__main__':
+    #! cache file
+    f = open("diff_images.json", 'r').read()
+    content = json.loads(f)
+    
     #! get config
     for image in ConfigYaml.images_cache:
         #! run docker command
         print("[GetImages] {}".format(image))
         saveImage=re.sub("@.*","",image["s_image"])
+        if saveImage in content:
+            print('Pass: {saveImage}')
+            continue
         docker_pull_cmd = "sudo docker pull {0}".format(image["t_image"])
         docker_tag_cmd = "sudo docker tag {0} {1}".format(image["t_image"], saveImage)
         docker_rm_cmd = "sudo docker rmi {0}".format(image["t_image"])
@@ -23,5 +31,8 @@ if __name__ == '__main__':
         )
         if ret == 0:
             print("[GetImagesDone] {}".format(image))
+            content.append(saveImage)
         else:
             print("[Error] {}".format(image))
+    with open("diff_images.json", 'w') as file:
+        file.write(content)
